+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog of lizard-nxt client
 Unreleased (2.0.0) (XXXX-XX-XX)
 ---------------------
 
+- Add region aggregation as a fourth aggregation tool. Draw regions and get
+  raster aggregations when clicked.
+
 - Add angular-gettext for translations.
 
 - Add grunt tasks to extract and compile translations.